docs: update readme with props
