+++ src/switch/colors.ts
@@ -1,0 +1,7 @@
+const Colors = {
+  activeTrackColor: "rgba(255,255,255,0.6)",
+  inactiveTrackColor: "rgba(0,0,0,0.2)",
+  white: "#ffffff",
+};
+
+export default Colors;
